+++ modules/Import.js
@@ -1,0 +1,7 @@
+export const navLink = document.querySelectorAll('.nav');
+export const contact = document.querySelector('.contact');
+export const bookSection = document.querySelector('.display-area');
+export const addBook = document.querySelector('.add-book');
+export const addToBook = document.querySelector('.book-form');
+export const displayContainer = document.querySelector('.book-display-container');
+export const Time = document.querySelector('.time');
